删除没用的第三方 密码功能bug修复（关掉后 就不用输入密码了） 优化 左滑删除按钮

+++ NoteBook/NoteBook/Sylar/Helper/CommomTools/CommonTools.m
@@ -33,7 +33,7 @@
 
 + (void) AutoLayoutVerticleView:(UIView*)_parentView WithSubviews:(NSArray*)_subViews TopEdge:(float)_topEdge BottomEdge:(float)_bottomEdge NeedCenter:(BOOL)_center
 {
-    int subview_count = [_subViews count];
+    NSInteger subview_count = [_subViews count];
     if (subview_count > 0)
     {
         float total_subview_height = _topEdge+_bottomEdge;
